requirements.txt
telethon
fastapi
uvicorn
python-telegram-bot
psutil
